pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pymongo = "^4.8.0"
httpx = "^0.27.2"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "-q", 
    "--tb=short",
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*",
    "ignore::PendingDeprecationWarning:starlette.*", 
    "ignore::DeprecationWarning:crypt.*",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
